+++ src/js/VertTuple.d.ts
@@ -165,7 +165,7 @@ export declare class VertTuple<T extends VertTuple<T>> {
      * @instance
      * @memberof VertTuple
      **/
-    getClosestPoint(p: Vertex): Vertex;
+    getClosestPoint(p: Vertex, constrain: boolean): Vertex;
     /**
      * The the minimal distance between this line and the specified point.
      *